+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `init` section is now `load` with fields `database` and corpus, saving the graph in its native format is done in a separate `save` section
   with a single target path. The section appears after the `export` and before the `footer` section.
+- `revise` updates after each modification of the graph
 
 ## [0.37.1] - 2025-07-23